testado no circuito 1 versao

+++ luminaria.cpp
@@ -1,6 +1,6 @@
 #include"luminaria.h"
 
-int luminaria::leitura_tensao()
+float luminaria::leitura_tensao()
 {
 
   //lê uma amostragem a cada 5° durante um período
@@ -19,7 +19,8 @@ int luminaria::leitura_tensao()
     }
   }
 
-  maior_Valor = maior_Valor * 0.001220703;    // (0,4096)->(0,5)
+  maior_Valor = maior_Valor * 0.001220703;    // (0,4095)->(0,5)  STM32
+  //maior_Valor = maior_Valor * 0.00489;      // (0,1023)->(0,5)  Arduino Uno
   valor_Corrente = maior_Valor - 2.5;         // diminui o valor de offset
   valor_Corrente = valor_Corrente*20;         // (0,1V)->(0,20A)
   valor_Corrente = valor_Corrente/1.41421356; // Vp->Vrms
