+++ luminaria.h
@@ -7,12 +7,13 @@
 class luminaria
 {
   public:
-         int leitura_tensao();
+         float leitura_tensao();
 
   private:
-          float vetCorrente[720];
-          double maior_Valor = 0;
-          double valor_Corrente = 0;
+          float vetCorrente[720];     // STM32
+          //float vetCorrente[320];   // Arduino Uno
+          float maior_Valor = 0;
+          float valor_Corrente = 0;
 };
 // ***  realiza a leitura da tensão analógica e retorna o valor da corrente;
